Collect warnings and output them at the end
Introduced a new setting for `WARN_AS_ERROR` i.e. `FAIL_ON_WARNINGS_PRINT`.
Especially on CI systems when running large projects it can be hard to obtain all the warnings as either they are written to a file and the file is not accessible (anymore) or they are written between the progress messages.
With this new option the warnings are given always at the end of the run.