Fix enumvalueType.name always None
This is a demonstration of how the generated code in compound.py is broken.  The generated code is fairly impenetrable and I am not familiar with the tool that produced this code from compound.xsd, however from the context of the other types in the file, I'm fairly sure that this is the code that is "supposed" to be here.  I don't see any problem with compound.xsd that should cause this error, so I'm assuming this is a bug with the code generator.  I have not done any testing or review to see if this problem exists with any of the other types, however my own application (which makes extensive use of doxmlparser) does not experience any other bugs of this type.