Prevent list items in section titles
When having the, a bit unusual, construct:
```
\section ThePage_subsection 1. The First Sub-Section
```
and we use the reference like:
```
\ref ThePage_subsection
```
we get in the LaTeX generation the error:
```
! LaTeX Error: Something's wrong--perhaps a missing \item.

See the LaTeX manual or LaTeX Companion for explanation.
Type  H <return>  for immediate help.
 ...

l.4 \end{DoxyEnumerate}}}
```
due to the construct:
```
Check it here \mbox{\hyperlink{_the_page_ThePage_subsection}{
\begin{DoxyEnumerate}
\item The First Sub-\/\+Section
\end{DoxyEnumerate}}}
```

This is all because of the `1.` i.e. a list item in the section title.

(all based on the discussion in:  Doxygen regression when creating PDF output 1.9.6 -> master of 10.01.2023 #9781 )