Fix enumvalueType
This is a potential solution to the issue with doxmlparser.  I don't think enumvalueType is supposed to be a mixed type.  The subelements are definitely supposed to be mixed types (linkedTextType and descriptionType).  But enumvalueType?  Without being familiar with the internals of Doxygen I can't say for sure.  But looking at the XML output it doesn't seem to contain mixed content.