Image appears in the middle of code in the internal documentation
Due to the fact that doxygen searches for a replacement string and this string is also part of the doxygen code we see in the internal documentation the weird behavior that in the middle of some source code of the description there is an image (see https://doxygen.github.io/doxygen/db/d14/index_8cpp.html and search for 3694). By splitting the string this can be overcome.