Not documented parameter in case of `const <type>` without argument name
When having an argument in a function definition like `const item` without an argument name and the type is `item` doxygen will see the word `item` as the name of the argument (unless the type is a predefined type from the standard like `int` of `std::size_t`).
When the type is found so far is just `const` we should see what is now in `name` as type and process further.

Example:
```
/// a fie
/// @param b2 the param
/// @return the return
int t_c_item2(int b2, const item);
```
giving the warning:
```
bb.h:15: warning: The following parameter of t_c_item2(int b2, const item) is not documented:
  parameter 'item'
```

(same applies for volatile, seen also the code just above)