Handling debug information in in respect to mscgen
By default there would be debug information in case of a debug build. In the other lexers used by doxygen the information is compiled into the debug executable but only shown upon request (`-d lex:<lexername>`).
The result is that when a msc image is generated the debug information will appear and disrupts the output.
For the mscgen lexer there is no such mechanism as it is an external lexer, building it into this lexer would make the lexer dependent on e.g. the src directory etc. and this is not wanted.