issue #9861 Links among markdown files don't work if upstream path element contains space
Doxygen replaces the given path / filename with the full path and due the the spaces in the path this leads to "problems" during further processing.
The problem does not only occur due to the "upstream" part but also to local path / filenames with spaces.
By embedding the link name always in double quotes (`"`) the problem can be overcome when the double quotes are handled..