Limit index fields in hhk file of chm file.
Based on some comments in #9894 where a warning was shown like
```
Warning: Keyword string:
...
is too long.  The maximum size is 488 characters.
```
tit was found that this was due to an extremely long `value` field in the `index.hhk` (index file), limiting this field and placing an ellipse when necessary