Unknown generated markdown id
Due to the initial fix the fix for #9254 didn't work anymore.
The problem is that there are 2 types of usage here:
- the code used for the situation as in #9254:
  ```
          docs = docs.left(match[1].position())+               // part before label
                 newLabel+                                     // new label
                 match[2].str()+                               // part between orgLabel and \n
                 "\\ilinebr @ianchor{" + orgTitle + "} "+orgLabel+"\n"+           // add original anchor plus \n of above
                 docs.right(docs.length()-match.length());     // add remainder of docs
  ```
- the code used with "@page" replacement like:
  ```
  docs.prepend("@ianchor{" + title + "} " +  autoId + "\\ilinebr ");
  ```

the subtle difference is the `\\ilinebr" and this would flow (plus the part directly after it) into the id as used by the `@ianchor` resulting in that there is no match.