Remove obsolete parts from translator.py
In the past it was possible only to select a number of output languages, nowadays all output languages are automatically included.
In the report generation etc. part there were still some parts that were that supported his mechanism but perturbated the flow in the script.
The obsolete parts have been removed