bug 630661 Highlighting lines in a dot graph
Based on the method as used by plantuml when `!pragma svginteractive true`  is set in doxygen a similar implementation is made:
- when hovering over a node all nodes and edges not directly connected to this node are blurred
- code used `svg.min.js` and loaded it from the internet, the file has been downloaded (so doxygen can still be used offline)  and adjusted (removed the reference line `//# sourceMappingURL=svg.min.js.map`)

To accomplish this we need unique edge identifications (hence a counting of the edges per graph)  and also the node identifications have to be in a way that a node id cannot be an abbreviation of another (e.g. `Node1` is an abbreviation of `Node10`).