Undefined references in pdf documents
In the doxygen test suite we see warnings like:
```
LaTeX Warning: Reference `013__class_8h_source' on page 3 undefined on input line 3.
```
which result in `??` in the resulting pdf file.
(looks like to happen mostly occurs for source files without file description).

By splitting the label and index part (analogous to what is done in rtfgen.cpp) the problem is resolved.