Detection of missing / superfluous `\n` in warnings
A number of warning messages have a double '\n' due to the fact that the message itself contains a '\n' and the warning routine automatically adds a '\n'.
Some messages are lacking a `\n` at the end which should be present.
In #6924 an automatic correction was attempted but a better solution would be that there is a compile time check that detects the missing / superfluous '\n'.
It is hardly possible to find all automatically as one can always trick a bit by means of something like `err("text %s",str) ` where `str` ends with a `\n`.

Some special constructs were necessary to overcome problems with macro substitution
- tagreader.cpp now using `p_warn` for internal / private warn function
- template.h special constructs, disabling `warn` macro with checks (checks done visibly) and directly calling function.
- trace.h, uses spdlog that has an own `warn` function that is temporary disabled.
- vhdlparser.jj has a function `term`, based on the name `term` in the VHDL standard which has nothing to do wit the doxygen `term` function and is renamed.