Interleave of `stdout` and `stderr`
During testing of #10129 we saw that there were some warnings that should have been at the end of log but appeared in the middle of the (last) lock of `Patching ...`.
This is due to the fact that `stdout` and `stderr` are less good synchronized for non terminal modes. We should flush the `stdout` before writing the block of warnings