Add 'extern' to xml output
Following the strategy taken in the sqlite output, we add extern if it
is set regardless of whether it is a function, variable, etc.

It might be reasonable to only do it for functions and variables
though. I don't know what the standard approach for these things is
in the code base.

We have to update addGlobalFunction so that it passes on the
'explicitExternal' information from Entry to the MemberDef which
previously was only done for variables and enums.

Also:

- Update the xsd file to indicate that an optional 'extern' attribute
  might be present on the members.