Show not up to date message  for translator
In:
```
Commit: 4c47b238edf4f94894ff70f6b32904e142f159a7 [4c47b23]
Date: Saturday, February 26, 2022 10:44:11 PM
Downgrade stale translator warning to message
```
The text was downgraded to a message and in this way dependent on the setting of `QUIET` whilst other configuration settings are given.
Also the text was written to `stdout` instead of `stderr`.