Wrong links for stateToString in internal doxygen documentation
When looking at the code of e.g. scanner.l in the doxygen internal documentation and looking at the prototype of the function `stateToString` this function is linked to the same function in the fortranscanner (although this should no happen as the function is `static`).
Due to the fact that the `stateToString` functionality is only enabled in the fortranscanner, but in the other lex scanners the prototype and the implementation is hidden inside a `#if` block but the (lex)code scanners still find it as it does not take the `#if` into account.
- enable the functionality of `stateToString` everywhere even when not used
- make the prototypes `[[maybe_unused]]` to satisfy the compilers (warnings)
- adjust `pre.l` so that the counting of the open / close brackets is correct when they are used in the rules (code taken from lexscanner.l)
- some minor fixes to keep the preprocessor happy  (commands in regular comments, escape double quotes in `[...]` part of the rules