Properties in python "namespaces"
When having in a python statement like (not in a class or function):
```
cached_property0 = property
```
this will give a warning like:
```
error: NamespaceDefImpl::insertMembers(): member 'cached_property0' with unexpected type id 10 and class scope '' inserted in namespace scope 'resolver'!
```

This is due to the tact that this is in python a property, see rule (in pyscanner.l):
```
^{B}{IDENTIFIER}/{B}"="{B}"property" { // property
```

this has  been corrected specifically for python (also necessary in filedef as otherwise the property is shown as variable).