Handling of relative include path in preprocessor
In the 7z package we get a warning like:
```
error: NamespaceDefImpl::insertMembers(): member 'CCoderReleaser' with unexpected type id 8 and class scope '' inserted in namespace scope 'NCompress::NArj::NDecoder'!
```
this because of the fact that the include files like:
```
#include "../../../C/CpuArch.h"
```
are "not found" in the preprocessor and thus a number of macros are not evaluated (7z heavily relies on macros), so we check now the absolute path and the path as done before so the `INCLUDE_PATH` still can be applied.