Handling comment in macro name
After fixing the include path problem in the 7z package (see PR #10383) a warning like remained / surfaced:
```
CPP/7zip/Archive/7z/7zExtract.cpp:24: warning: Found ';' while parsing initializer list! (doxygen could be confused by a macro call without semicolon)
```
this is due to code (strongly abbreviated and modified) like:
```
/// \file

#define A(x,y) x ## y(int i)
#define fie1(x) FIE1(x);

A(fie,1 /* test */)

A(fie,1)
```
which gives in the preprocessor:
```
00001 /// \file
00002
00003 #define A(x,y)
00004 #define fie1(x)
00005
00006  fie1 /* test */ (int i)
00007
00008  FIE1( int i );
00009
```
whilst we would expect also on line 6
```
00006  FIE1( int i );
```
yjis is due to the comment between the name of the (after the first pass) macro and the arguments.