Handling of `import` in non cpp files
Based on the question: https://stackoverflow.com/questions/77366144/is-there-a-way-to-configure-doxygen-such-that-references-between-packages-also-b?noredirect=1#comment136392319_77366144
In the doxygen version 1.9.7 the `import` satement was working for e.g. the Java language but this is not the case anymore for the version 1.9.8 due to the fact that the handling is geared towards cpp `module`s.
In case not in Cpp we now revert back to the old code.