Fix @enum command with @ingroup
Since commit bd99649d9d2d15cb3c1c7b53167365b79a6aaa5b the groups of an
enum documentation entry are used to associate the documentation entry
with a definition.  However, this broke the case were a matching
definition is not yet associated with a group.  Reorder the searches to
fix this issue.

Fix #10151.