Line miscounting in case of structural indicator
In case the (first) structural indicator is not at the beginning of a comment block, the line counting is not correct and we get warnings like:
```
.../dd.h:4: warning: Found unknown command '@error_2'
.../dd.h:8: warning: Found unknown command '@error_6'
.../PrivacyideaService.php:22: warning: Unsupported xml/html tag <service> found
.../PrivacyideaService.php:29: warning: Found unknown command '@subpackage'
```
whilst the correct line numbers should be here: 2, 6, 13 and 20.

The problem only occurs for the first structural indicator.
(It is a bad idea to have the, first, structural indicator in the middle of a comment block.)