Incorrect warning and output with `\image` followed by HTML command
When having:
```
@image html PCore-Block-Diagram.png width=100px <b>bold</b>
```
we get a warning like:
```
dox_main.md:16: warning: found </b> tag without matching <b>
```

and in the output we see as running text: `bbold`

When having instead of a HTML command a doxygen command or a symbol (`&...;`) we don't get a warning.
The HTML command should be seen as end of the underlying command and not be handled at this moment (not the code is also used on other places hence the conditional).