Incorrect setting return state after `\include{doc}`
When having:
```
/// \file
///
/// The level 1
///
/// \include kaboom.md kaboom.md
///
/// \include{doc} kaboom.md kaboom.md

/// the fie 1
void foo_1();

/// the class 1
class cls_1{};
```
we get the warning
```
.../foo.h:20: warning: Member foo_1() (function) of file foo.h is not documented.
```
and we see in the resulting documentation that the content from `kaboom.md` is twice present present for `\include{doc}` but only once for `\include` togehtehr wit the extra text `kaboom.h`.

This problem is due to the fact that the old jump back state was set to `YY_START` which is the state `IncludeDoc` (from rule `<IncludeDoc>{FILEMASK}|"\""[^\n\"]+"\"" {`) but should have been the state `IncludeFile` from the rule `<CComment,ReadLine,IncludeFile>[\\@]("include"{B}*"{doc}"|"includedoc") {` (and properly stared in yyextra->includeCtx).