Incorrect jump to anchor inside page when using CTRL-click
When having a simple example like:
```
# Hello World

[TOC]

## Introduction
This is the canonical `Hello World` program.

## Reference
This is a second paragraph
```
and clicking in the treeview on "Hello World" followed by a click on "Introduction" a jump to "Introduction is performed but when doing a CTRL_click on "Introduction"  (so opening a new tab) we get as page "javascript:void(0)"  which is empty. In case of a jump "inside" the page we should jump to the anchor.