issue #10516 @copybrief command does not take effect with 1.10.0
For better error / warning handling a line with file / line information was added to the brief description. When a new brief line arrives it was checked whether or not it could be the brief description by checking against the empty string, the later was not the case anymore and hence such a line should not be taken into consideration for determining the empty-ness of the original brief description.