Warning about invalid argument with `\iline` command
When having a simple file like:
```
/**
 * \if english
 * \brief  Namespace
 * \endif
 *
 * \if german
 * \brief  Namensraum
 *
 * \endif
 */
namespace dBaseRelease
{
}
```
with
```
ENABLED_SECTIONS       = german
```

We get the warning like:
```
.../BoostServer.cc:6: warning: invalid argument for command '\iline'
```

Initial the `iline` command is, when inserted, properly inserted but due to the call of `stripTrailingWhiteSpace` (in commentscan.l) at that moment trailing whitespace and line breaks are removed.
The `iline` value can now also be terminated by means of the start of a command.