Rendering of general `dl.section` styles
In:
```
Commit: e24d8a0589956032d30f53c85f6f9693c55a1a70 [e24d8a0]
Date: Sunday, November 19, 2023 10:48:26 AM
Improved rendering of sections like @note, @warning, @todo etc. for HTML
```
the rendering of a number of sections was improved, but this should only be done for sections that have a specific background and not for the general `dl.section`.
The rendering has now been reversed for the general `dl.section` and specific sections are listed to be rendered in the "new" style.