Incorrect line number and filename after `\include{doc}` and `\snippet{doc}`
After the commands `\include{doc}` and `\snippet{doc}` the information in warnings from the comment scanner points to the includer file and the line numbers keep increasing even when the warning comes from code in the included file. This is due to the fact that the commands are already handled in the comment converter and the, added, `\ifile` and `\iline` commands are not taken into account in the comment scanner.
Also the line number from a `\snippet{doc}` command is not correct, it doesn't refer  to the line of the snippet but to the beginning of the file (due to the hard coded `\iline 1`).