Better warning when in Fortran a comment block is not closed
When having e.g. a stray `@code` in the comment we get a warning like:
```
Error in file .../gmsh-4.12.2-source/api/gmsh.f90 line: 15475, state: 29(DocCopyBlock)
```
where the line 15475 is the end of the file.
It is a bit hard to find where the problem actually occurs, so a warning like:
```
Error in file .../gmsh-4.12.2-source/api/gmsh.f90 line: 15475, state: 29(DocCopyBlock), starting command: '@code' probable line reference: 6689
```
is better