Incorrect line number and filename after `\include{doc}` and `\snippet{doc}`
We need the `ifile` and `iline` command to be `Replace`d in case of a section header as otherwise we get still incorrect filename / line numbers.
(In case the `ifile` / `iline` is added by an `include{doc}` or a `snippet{doc}` command everything worked already but as seen as the `ilinebr` is gone there are some problems.)

We need in `commentcnv.l` the extra space as otherwise we run in the `commentscan.l` into the rule (and also add it in `docparser.cpp` for consistency) .
```
<Comment>("\\"[a-z_A-Z]+)+"\\"
```
which "eats" the `\` from `\ilinebr\ifile`.