Fix table overlaping footer
This bug lies in Latex longtable since ages. doxygen, and  pandoc,
suffer this bug since. See #6532.

The issue is that a table may overlap the footer. There are even cases
where the bottom of a table is cropped at the end of the page.

The culprit is the longtable latex package. This bug is known for a long
time, and a patch has been proposed since 2003. This package is not
maintained anymore.

- https://github.com/jgm/pandoc/issues/2889#issuecomment-245879494
- https://www.latex-project.org/cgi-bin/ltxbugs2html?pr=tools/3512

The proposed patch has been applied without any conflict.

The result is that there is not more table overlap. See example in pull
request #10667.

Fixes #6532