Handling of empty structural command
When having an empty structural command like `\var` or `\namespace` we get different types of warnings and the information following is lost.

Example:
**aa.h**
```
/// \file

/**
 * @var
 *
 * \brief MY_var var1
*/
int var1  = 42;
/**
 * @var var2
 *
 * \brief MY_var var2
 */
int var2  = 42;
```

**bb.h**
```
/// \file

/**
 * @namespace
 *
 * \brief MY_namespace ns1
*/
namespace ns1 {}
/**
 *
 * @namespace ns2
 *
 * \brief MY_namespace ns2
 */
namespace ns2 {}
```
results in the warnings:
```
.../aa.h:8: warning: Empty prototype found!
.../bb.h:6: warning: missing argument after \namespace.
.../aa.h:5: warning: member with no name found.
.../aa.h:9: warning: Member var1 (variable) of file aa.h is not documented.
```
and the information of the "empty" structural comment blocks is lost,
whilst expected would be:
```
.../aa.h:5: warning: missing argument after '\var'.
.../bb.h:6: warning: missing argument after '\namespace'.
```
and the information would be assigned to the next  entity.

(we need the `return checkStructuralIndicator(yyscanner);` as otherwise when there are multiple structural commands in a comment block they are no handled properly).