Missing items in doxygen  internal search
When having the simple file:
```
\mainpage The main page

\page pg1 The first page

\section sect_1_1 first section first page

\section sect_1_2 second section first page

\page pg2 The second page

\section sect_2_1 first section second page

\section sect_2_2 second section second page
```
and doing a "normal" internal doxygen search for the word `second` we get a.o.
```
first section second page
The second page
second section first page
second section second page
```
but when searching for `second page` we get
```
first section second page
The second page
```
so we are missing the reference to
```
second section second page
```
this is due to the filtering of the words in the string and omitting double words.