Curly brackets in options
When  having a markdown file like:
```

```
we get warnings like:
```
warning: Unexpected character '}' while looking for section label or title
```
due to the fact that the `##### mybytestring {bytestring}` heading is translated to
```
\ianchor{mybytestring {bytestring}} autotoc_md2\ilinebr <h5>mybytestring {bytestring}</h5>
```
later on (commentscan) the fist `}` is seen as end of the `\ianchor` option.

By temporary changing the `}` this is prevented.