Unique labels for `\include{doc}`
With the `raise` option with the `\internal{doc}` command it was possible to raise the level of the used sections to the sub-sections specified.
Problem would still be that the same `label_id` would have been used which would give warnings like:
```
 warning: multiple use of section label 'cmd_1' while adding section
```
by creating also the `label` option this can be prevented, the specified value will be, recursively, appended to the given `label_id`.
This would not only happen for section ids but also for anchor ids.

- option handling in commentcnv
  - more flexibility in the order in which the options can be used (only a fixed order was possible so far).
  - add handling of the new option `label`
- comentscan
  handling of the `label` with the `\iraise` command
- updating documentation
  - for the new option
  - making `\includedoc` also have the possibility to use the `raise` and `label` options.
- updating the internal documentation of `\iraise` to accommodate the `label_id`.

(everything also applies the `\snippet` type of commands)