Automatic update of the addon `py` files based on `xsd` file
When a new feature is implemented it is possible that this has consequences for the `compound.xsd` or `index.xsd` files and thus also for the derived `py` files in the doxmlparser addon.
The later is easy to oversee as the update of the `xsd` files is often noticed during the tests and with the `make tests` the update of the `py` files was not performed.
- adding the dependency to the `tests`
- removing the possibility to optional adding the update of the doxmlparser (the update was a reminiscence of the time that the doxmlparser was still a cpp file).