Flexibility for trDocumentation  in respect to the ProjectName
In the proposed pull request #439 an initial fix was made for the fact that some languages have a different order of the words when appended.
This is not only the case for right to left languages but also for e.g. Spanish, Brazilian  (see comments in their translators).

This implementation gives the flexibility to the respective translators in respect to the wording.