Warnings due to multiple "sectioning" commands inside an `\if` type construct
When having a construct like:
```
/*!
@if LANG_ENGLISH
@file test.h

This is the description of test.h

@class Dtk::Core::Fn_Test test.h
@brief Fn_Test class.

Details about Fn_Test.

@fn const char *Fn_Test::member(char c,int n)
@brief A member function.

Details of the function

@param c a character.
@param n an integer.
@exception std::out_of_range parameter is out of range.
@return a character pointer.

@fn void Fn_Test::undocumented()
@brief test

@endif
*/
```
we get warnings like:
```
.../test.en_US.dox:7: warning: Documentation block ended in the middle of a conditional section!
.../test.en_US.dox:25: warning: found \endif without matching start command
```
due to the fact that the part about the `\file` is ended by means of the `\class` command and later restarted but the test about the `\if` termination does not know about this restart (and thus the warning), and when restarted the `\if...` stack is newly defined specified so the `\endif` is on its own.

We need to "save" the stack for further use.