Unicode characters in file names
When having a markdown file like:
```
# Test for having images with non-ASCII characters in their name

![Kartöffelchen](kartöffelchen.jpg)
```
we get the warnings like:
```
aa.md:3: warning: image file kart is not found in IMAGE_PATH: assuming external image.
```
(file is present and IMAGE_PATH is correctly set).
This is due to the fact that unicode characters are not supported in all lexers for file names.
In `commentscan.l`  they are supported, adding them to `doctokenizer.l` would solve this issue, but to be consistent the characters are also added to `*scanner.l`