Code sections in Perl documentation not expandable anymore
In the Perl helper function (`doxygen-filter-perl`) for a long time already the function `toggleVisibility` was used to show the dedicated inline, collapsible, code.
As of doxygen 1.10.o this doesn't work anymore as the `toggleVisibilty` is now in a `let` construct and should be used as `dynsection.toggleVisibility`.
For compatibility adding a function `toggleVisibility` that calls `dynsection.toggleVisibility`.