Warning in case of `0.` as start of a list.
When having an example like:
```
\mainpage The mainpage

Steps:

0. First step
  - sub step
1. Second step
```
we get a warnings like:
```
aa.md:7: warning: Invalid list item found
```
this is due to the fact that `0.` is not recognized as start of a list.

There is no objection against having `0.` as start of a list, added it.

When not having the sub element the `0. ...`  is seen just as regular text.

(Note due to the documented way doxygen handles the list numbering, the `0.` is shown as `1.` and the original `1.` as `2.).