Setup for doxygen documentation
In #11006 we saw problem regarding the usage of jquery.js in CHM but there were in the doxygen documentation CHM file also problems regarding the clipboard.js and menu.js (due to the incorrect setting of `HTML_COPY_CLIPBOARD` and `HTML_DYNAMIC_MENUS`  (only one set of examples was generated and this was geared towards the "normal" html documenaton).
Now 2 sets of examples are generated (one for the "normal" html documentation and one for the chm documentation) and furthermore the stylesheet usage in the examples has been mad compatible with the doxygen documentation and in the examples for the "normal" html documentation the `HTML_COLORSTYLE=TOGGLE` is used to be consistent with the main documentation.