Compilation error under Cygwin on language.cpp: "too many sections"
In case we build under Cygwin under debug mode we get for language.cpp the message about "too many sections".
This can be prevented by means of using the extra : `-Wa,-mbig-obj`.

(Might be that this problem also occurs under MinGW but due to missing test environment it is not added there yet)