Fix out-of-bounds string_view access in Markdown n-/m-dash processing
When Doxygen is built with libstdc++ assertions enabled (passing
`-Wp,-D_GLIBCXX_ASSERTIONS` to the compiler, which is for example how
the ArchLinux doxygen package is built), the markdown processing code
can hit an OOB condition when processing en- and em-dashes. Which then
manifests as a quite nasty abort:

    /usr/include/c++/14.2.1/string_view:256: constexpr const std::basic_string_view<_CharT, _Traits>::value_type& std::basic_string_view<_CharT, _Traits>::operator[](size_type) const [with _CharT = char; _Traits = std::char_traits<char>; const_reference = const char&; size_type = long unsigned int]: Assertion '__pos < this->_M_len' failed.

Added a minimal reproduction case to the markdown test and updated the
expected output to verify this doesn't regress again in the future.