Output when brief description ends with multiple spaces
When we have multiple spaces at the end of a brief description like in:
```
#define BRLAPI_PACKET_RESUMEDRIVER0    'R'   /**< Resume driver               */
```
and settings:
```
JAVADOC_AUTOBRIEF = YES
MARKDOWN_SUPPORT = YES
GENERATE_MAN           = YES
```
This leads in e.g. the HTML output to some extra white space for the brief description and in the MAN output to:
```
       #define BRLAPI_PACKET_RESUMEDRIVER0   'R'
           Resume driver
            "
```

The problem is the empty spaces at the end of the brief description that are converted ny the `MARKDOWN_SUPPORT` to `\ilinbr<br>`, the brief description is now stripped of leading and trailing whitespce and also the later is now stripped.