Remove warnings regarding bison package
When running bison we get warnings like:
```
Sebastien Loriot

9:07 AM (5 hours ago)

to me, Laurent
Hello,

today's testsuite is fully red because of a python issue that has
nothing to do with doxygen.

While writing to you, here are some warnings you might want to fix:

/home/cgal-testsuite/cgal_doc_build/CGAL-6.0.1-Ic-349/doc/scripts/doxygen_master/src/constexp.y:36.1-25:
warning: deprecated directive: ‘%name-prefix "constexpYY"’, use ‘%define
api.prefix {constexpYY}’ [-Wdeprecated]
    36 | %name-prefix "constexpYY"
       | ^~~~~~~~~~~~~~~~~~~~~~~~~
       | %define api.prefix {constexpYY}
/home/cgal-testsuite/cgal_doc_build/CGAL-6.0.1-Ic-349/doc/scripts/doxygen_master/src/constexp.y:36.1-25:
warning: deprecated directive: ‘%name-prefix "constexpYY"’, use ‘%define
api.prefix {constexpYY}’ [-Wdeprecated]
    36 | %name-prefix "constexpYY"
       | ^~~~~~~~~~~~~~~~~~~~~~~~~
       | %define api.prefix {constexpYY}
/home/cgal-testsuite/cgal_doc_build/CGAL-6.0.1-Ic-349/doc/scripts/doxygen_master/src/constexp.y:
```

The problem was called because of the usage of an old bison version (2.7) in GitHub Actions for windows so the `%name-prefix "constexpYY"` was necessary.
There is a newer version of bison available on chocolaty and also on sourceforge so no reason for using the old version.